requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.13.1
django-extensions==3.0.2
graphviz==0.14
mysql-connector-python==8.0.20
protobuf==3.12.2
pydot==1.4.1
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
